% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_open_outlines_from_closed_outlines.R
\name{get_open_outlines_from_closed_outlines}
\alias{get_open_outlines_from_closed_outlines}
\title{A function to create open outlines from closed outlines.}
\usage{
get_open_outlines_from_closed_outlines(
  outlines_combined,
  return_combined_outlines = TRUE
)
}
\arguments{
\item{outlines_combined}{A Momocs Out file containing closed outlines.}

\item{return_combined_outlines}{(default = TRUE) A logical parameter
stating whether to output a Momocs Opn-file, or a list of coordinate
matrices of each open outline.}
}
\value{
If return_combined_outlines = TRUE, returns the combined Coo
objects in a single Opn file. If return_combined_outlines = FALSE, returns
a list of coordinate matrices of each open outline.
}
\description{
The way it is implemented, the open outline will start at the highest,
left-most coordinate, run clock-wise, and will end at the lowest,
left-most coordinate.
}
