% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outlines_function.R
\name{get_outlines}
\alias{get_outlines}
\title{get_outlines}
\usage{
get_outlines(outpath, tps_file_rescale = NULL)
}
\arguments{
\item{outpath}{The path to the folder containing the .jpg files of the singled out artefacts
from which the outlines should be derived (as for example created in the
separate_single_artefacts function).}

\item{tps_file_rescale}{(Default = NULL) A dataframe containing at least a column \emph{IMAGE} and a column \emph{SCALE}.
This dataframe is internally created using the function \emph{tps_to_df} (for more information see ?tps_to_df()).
\emph{IMAGE} contains the names of the images, as they are written in \emph{inpath}.
\emph{SCALE} contains the scaling factor from pixel to a metric measurement.
Such a file can be created in i.e. tpsDIG2 (Rohlf 2017).}
}
\value{
Returns a Momocs Out file containing the outlines' coordinates
with their associated IDs, derived from the file name.
}
\description{
Uses Momocs::import_jpg() to import outline coordinates from the single
.jpg files which are either already avaiable, or were created using the
separate_single_artefacts function. This function preserves the
filenames.
}
