% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outlines_function.R
\name{get_outlines_function}
\alias{get_outlines_function}
\title{Uses Momocs::import_jpg1() to import outline coordinates from the single .jpg files which are either already avaiable, or were created using the separate_single_artefacts_function function. This function preserves the filenames.}
\usage{
get_outlines_function(pathname_output)
}
\arguments{
\item{pathname_output}{The path to the single .jpg files of the artefacts from which the outlines should be derived (as for example created in the separate_single_artefacts_function function).}
}
\value{
Returns a Momocs Out file containing the outlines' coordinates with their associated IDs, derived from the file name.
}
\description{
Uses Momocs::import_jpg1() to import outline coordinates from the single .jpg files which are either already avaiable, or were created using the separate_single_artefacts_function function. This function preserves the filenames.
}
