% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_outlines_from_closed_outlines_v3.R
\name{open_outlines_from_closed_outlines}
\alias{open_outlines_from_closed_outlines}
\title{open_outlines_from_closed_outlines}
\usage{
open_outlines_from_closed_outlines(
  outlines_combined,
  return_combined_outlines = TRUE
)
}
\arguments{
\item{outlines_combined}{A Momocs Out file containing closed outlines.}

\item{return_combined_outlines}{(default = TRUE) A logical parameter stating wether to output a Momocs Opn-file, or a list of coordinate matrices of each open outline.}
}
\value{
If return_combined_outlines = TRUE, returns the combined Coo objects in a single Opn file. If return_combined_outlines = FALSE, returns a list of coordinate matrices of each open outline.
}
\description{
A function to create open outlines from closed outlines.
The way it is implemented, the open outline will start at the highest, left-most coordinate,
run clock-wise, and will end at the lowest, left-most coordinate.
The starting point is found using the shortest distance from an arificial point on X = 0 and Y = max(current_outline_df$Y) (i.e., the overall highest Y-coordinate),
to the artefacts outline. The ending point is found using the starting points X-coordinate and it's corresponding minimum Y-coordinate.
}
\details{
This works only for artefacts, which have their supposed open side on their left and have been prepared in a way, so that this open left side is a straigt vertical line.
}
