% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_single_artefacts_function.R
\name{separate_single_artefacts}
\alias{separate_single_artefacts}
\title{separate_single_artefacts}
\usage{
separate_single_artefacts(inpath, outpath)
}
\arguments{
\item{inpath}{Pathname to the folder containing the JPEG
images with multiple artefacts on it.}

\item{outpath}{Pathname to the folder where the JPEGs of
the singled-out artefacts from this function should be stored.}
}
\value{
If return_combined_outlines = TRUE, returns the combined
Coo objects in a single Opn file. If return_combined_outlines = FALSE,
returns a list of coordinate matrices of each open outline.
}
\description{
A function to separate single artefacts from an image containing
multiple artefacts.
}
