% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_single_artefacts_function.R
\name{separate_single_artefacts_function}
\alias{separate_single_artefacts_function}
\title{A function to separate single artefacts from an image containing multiple artefacts.}
\usage{
separate_single_artefacts_function(pathname_input, pathname_output)
}
\arguments{
\item{pathname_input}{Pathname to the folder containing the JPEG images with multiple artefacts on it.}

\item{pathname_output}{Pathname to the folder where the JPEGs of the singled-out artefacts from this function should be stored.}
}
\value{
If return_combined_outlines = TRUE, returns the combined Coo objects in a single Opn file. If return_combined_outlines = FALSE, returns a list of coordinate matrices of each open outline.
}
\description{
A function to separate single artefacts from an image containing multiple artefacts.
}
