% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tps_to_dataframe_function.R
\name{tps_to_df}
\alias{tps_to_df}
\title{tps_to_df}
\usage{
tps_to_df(tps_file_path = tps_file_path)
}
\arguments{
\item{tps_file_path}{The path to the .tps file containing "LM", "IMAGE", "ID", and "SCALE" information (no more no less), as for example created using tpsUtil and tpsDig2 (Rohlf 2017).}
}
\value{
Returns a dataframe with "LM", "IMAGE", "ID", and "SCALE" as column names.
}
\description{
Regular Expression loop to get .tps files containing "LM", "IMAGE", "ID", and "SCALE" information (no more no less) into a dataframe,
which subsequently can be used in the function \emph{get_outlines()} for scaling.
}
