% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_single_artefacts_function.R
\name{separate_single_artefacts_function}
\alias{separate_single_artefacts_function}
\title{A function to create open outlines from closed outlines. The way it is implemented, the open outline will start at the highest, left-most coordinate, run clock-wise, and will end at the lowest, left-most coordinate.}
\usage{
separate_single_artefacts_function(pathname_input, pathname_output)
}
\arguments{
\item{single_outlines_list}{A Momocs Out file containing closed outlines.}

\item{return_combined_outlines}{(default = TRUE) A logical parameter stating wether to output a Momocs Opn-file, or a list of coordinate matrices of each open outline.}
}
\value{
If return_combined_outlines = TRUE, returns the combined Coo objects in a single Opn file. If return_combined_outlines = FALSE, returns a list of coordinate matrices of each open outline.
}
\description{
A function to create open outlines from closed outlines. The way it is implemented, the open outline will start at the highest, left-most coordinate, run clock-wise, and will end at the lowest, left-most coordinate.
}
\examples{
separate_single_artefacts_function(pathname_input = "path/to/input_files", pathname_output = "path/to/output_files")
}
